Extracted Notifiable into standalone file, tidied FormattableContent a bit
