+++ WordPress/Classes/Utility/FormattableContent/FormattableContentGroup.swift
@@ -1,5 +1,16 @@
 import Foundation
 
+extension FormattableContentGroup.Kind {
+    static let text = FormattableContentGroup.Kind("text")
+    static let image = FormattableContentGroup.Kind("image")
+    static let user = FormattableContentGroup.Kind("user")
+    static let comment = FormattableContentGroup.Kind("comment")
+    static let actions = FormattableContentGroup.Kind("actions")
+    static let subject = FormattableContentGroup.Kind("subject")
+    static let header = FormattableContentGroup.Kind("header")
+    static let footer = FormattableContentGroup.Kind("footer")
+}
+
 // MARK: - FormattableContentGroup: Adapter to match 1 View <> 1 BlockGroup
 //
 class FormattableContentGroup {
