+++ WordPress/Classes/ViewRelated/Notifications/FormattableContent/Groups/BodyContentGroup.swift
@@ -81,7 +81,7 @@ class BodyContentGroup: FormattableContentGroup {
         }
     }
 
-    public class func pingbackReadMoreGroup(for url: URL) -> FormattableContentGroup {
+    private class func pingbackReadMoreGroup(for url: URL) -> FormattableContentGroup {
         let text = NSLocalizedString("Read the source post", comment: "Displayed at the footer of a Pingback Notification.")
         let textRange = NSRange(location: 0, length: text.count)
         let zeroRange = NSRange(location: 0, length: 0)
