+++ WordPress/Classes/ViewRelated/Notifications/FormattableContent/Styles/SnippetsContentStyles.swift
@@ -1,5 +1,5 @@
 
-class SnipetsContentStyles: FormattableContentStyles {
+class SnippetsContentStyles: FormattableContentStyles {
     var attributes: [NSAttributedStringKey: Any] {
         return WPStyleGuide.Notifications.snippetRegularStyle
     }
